changed combination calculation to take large factorials into account

+++ CPP/Combination.h
@@ -9,7 +9,7 @@ namespace C
             int calculateCombination(int n, int r);
 
         private:
-            int fact(int x);
+            unsigned long long int fact(int x);
     };
 }
